major refactor of __init__ into different files to improve structure; fix github basedpyright CI
